refactor: made code more modular
